move endpoint group
