+++ resources/views/oauth/authorized-clients/index.blade.php
