+++ routes/web.php
@@ -163,8 +163,6 @@ Route::resource('groups', 'GroupsController', ['only' => ['show']]);
 
 Route::group(['prefix' => 'home'], function () {
     Route::group(['as' => 'account.', 'prefix' => 'account'], function () {
-        Route::get('tokens', 'Passport\TokensController@tokens');
-        Route::delete('revoke-client/{clientId}', 'Passport\TokensController@revokeClient');
         Route::get('edit', 'AccountController@edit')->name('edit');
         // Uploading file doesn't quite work with PUT/PATCH.
         // Reference: https://bugs.php.net/bug.php?id=55815
@@ -210,6 +208,10 @@ Route::group(['prefix' => 'home'], function () {
 });
 
 Route::get('legal/{page}', 'LegalController@show')->name('legal');
+
+Route::group(['as' => 'oauth.', 'prefix' => 'oauth', 'namespace' => 'OAuth'], function () {
+    Route::resource('authorized-clients', 'AuthorizedClientsController', ['only' => ['index', 'destroy']]);
+});
 
 Route::get('rankings/{mode?}/{type?}', 'RankingController@index')->name('rankings');
